Fix bug in when verification is called
Fixes a bug introduced by change 422821. In this change, the schema
validation was moved from the 'prepare' to 'validate' method in
'_clean_up_images'. This causes the validation to be performed before
the provision.json has been generated for a given node and so results
in a Nova Scheduler error as '[Errno 2] No such file or directory'.
Moving the function call back to 'prepare' fixes this problem.

Change-Id: I9128abb5dbc0579190323979f93691e21f4ef6b2