Work around update to oslo.db
The changes made in the 4.19.0 release of oslo.db to remove the
sqlite_db option break the bareon-ironic unit tests. Ironic
was patched to fix this issue in the 8.x series (Pike), but
the fix (commit d31d9848) has not yet been backported to earlier
releases of ironic. Since the unit tests install ironic 6.x
(stable/newton) they fail with oslo.db >= 4.19.0.

This change works around the problem by placing an upper limit
on the oslo.db version used when running the tests. When the fix
has been backported in ironic, we can remove this change
and bump the version of ironic installed in tox.ini so that
it includes the fix.

See bug #1676809 for more details.

Change-Id: I4795476f254ed88ee4b7391daf0afbee95bf192d